requirements.txt
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
chardet==3.0.4
charset-normalizer==3.3.2
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.0
daphne==4.0.0
decorator==4.4.2
dj-database-url==2.1.0
Django==5.0
django-ckeditor==6.7.1
django-heroku==0.3.1
django-js-asset==2.2.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
googletrans==3.0.0
gunicorn==21.2.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.1.5
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
imageio==2.34.0
imageio-ffmpeg==0.4.9
incremental==22.10.0
moviepy==1.0.3
numpy==1.26.3
opencv-python==4.9.0.80
packaging==23.2
Pillow==10.1.0
proglog==0.1.10
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==24.0.0
pytesseract==0.3.10
pytz==2023.3.post1
requests==2.31.0
rfc3986==1.5.0
service-identity==24.1.0
setuptools==69.0.3
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.2
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.2.0
whitenoise==6.6.0
zope.interface==6.1
